tighten up login ui
